Move some test mods into appropriate packages
