+++ src/test/java/net/minecraftforge/debug/fluid/FluidPlaceBlockEventTest.java
@@ -1,4 +1,23 @@
-package net.minecraftforge.debug;
+/*
+ * Minecraft Forge
+ * Copyright (c) 2018.
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation version 2.1
+ * of the License.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ */
+
+package net.minecraftforge.debug.fluid;
 
 import net.minecraft.block.BlockStone;
 import net.minecraft.init.Blocks;
@@ -18,7 +37,7 @@ import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
  *  - Prevent lava from setting surrounding blocks on fire.
  */
 @Mod(modid = "fluidplaceblocktest", name = "FluidPlaceBlockTest", version = "0.0.0", acceptableRemoteVersions = "*")
-public class FluidPlaceBlockTest
+public class FluidPlaceBlockEventTest
 {
     private static final boolean ENABLED = false;
 
@@ -26,7 +45,7 @@ public class FluidPlaceBlockTest
     public void preinit(FMLPreInitializationEvent event)
     {
         if (!ENABLED) return;
-        MinecraftForge.EVENT_BUS.register(FluidPlaceBlockTest.class);
+        MinecraftForge.EVENT_BUS.register(FluidPlaceBlockEventTest.class);
     }
 
     @SubscribeEvent @SuppressWarnings("unused")
