+++ src/test/java/net/minecraftforge/debug/world/ChunkWatchEventTest.java
@@ -1,4 +1,23 @@
-package net.minecraftforge.debug;
+/*
+ * Minecraft Forge
+ * Copyright (c) 2018.
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation version 2.1
+ * of the License.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ */
+
+package net.minecraftforge.debug.world;
 
 import org.apache.logging.log4j.Logger;
 
@@ -8,8 +27,8 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
-@Mod(modid = ChunkWatchWorldTest.MODID, name = "Chunk Watch World Test", version = "1.0", acceptableRemoteVersions = "*")
-public class ChunkWatchWorldTest
+@Mod(modid = ChunkWatchEventTest.MODID, name = "Chunk Watch Event Test", version = "1.0", acceptableRemoteVersions = "*")
+public class ChunkWatchEventTest
 {
     public static final String MODID = "chunkwatchworldtest";
 
@@ -23,7 +42,7 @@ public class ChunkWatchWorldTest
 
         if (ENABLED)
         {
-            MinecraftForge.EVENT_BUS.register(ChunkWatchWorldTest.class);
+            MinecraftForge.EVENT_BUS.register(ChunkWatchEventTest.class);
         }
     }
