+++ src/test/java/net/minecraftforge/debug/world/WorldProviderCanSleepAtTest.java
@@ -1,4 +1,23 @@
-package net.minecraftforge.debug;
+/*
+ * Minecraft Forge
+ * Copyright (c) 2018.
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation version 2.1
+ * of the License.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ */
+
+package net.minecraftforge.debug.world;
 
 import org.apache.logging.log4j.Logger;
 import net.minecraft.entity.player.EntityPlayer;
@@ -10,8 +29,8 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 
-@Mod(modid = CanSleepAtTest.MODID, name = "CanSleepAtTest", version = "0.0.0", acceptableRemoteVersions = "*")
-public class CanSleepAtTest
+@Mod(modid = WorldProviderCanSleepAtTest.MODID, name = "Can Sleep At Test", version = "0.0.0", acceptableRemoteVersions = "*")
+public class WorldProviderCanSleepAtTest
 {
     public static final String MODID = "can_sleep_at_test";
     public static final boolean ENABLED = false;
@@ -37,7 +56,7 @@ public class CanSleepAtTest
         @Override
         public DimensionType getDimensionType()
         {
-            return CanSleepAtTest.dimType;
+            return WorldProviderCanSleepAtTest.dimType;
         }
 
         @Override
